refactor: moved responsive elements to shared components directory
