+++ app/(components)/Shared/Responsive/Text.tsx
