+++ app/(components)/[study]/[semester]/SemesterSelection.tsx
@@ -1,6 +1,6 @@
 'use client'
 import Card from "@/components/Cards/Card"
-import { Text } from "@/components/ResponsiveTags/Text"
+import { Text } from "@/components/Shared/Responsive/Text"
 import { Context, createContext, useContext, useState } from "react"
 import Link from "next/link"
